Make map view fit show all points on initial load
Previously we had a fixed default zoom level, and had avoided using the
simple
fitBounds() method, because it zooms right in when we have a single
point.
Instead we now use fitBounds to get all points on the map, but if zoom
level is further than 15 we set zoom to 15. This means a single point
still shows zoomed out enough to give a rough view of the surrounding
city.