Rough beginnings of create post wiring
Starting to wire up the beginnings of the create post process.
This commit adds an event handler when the create post form is
submitted.
I'm using backbone.syphon to pull data out of the form and serialize it
in a
form we can use for models. I'm not doing any handling of exceptions yet
so all saves will fail.