Adding Button Mixins
I have taken the classes out of the HTML markup and used Sass mixins in the .scss files.

I had to "eye" the Login/Register and Post Detail view, since they are not set up in the app.

I need to DOUBLE CHECK these when merged!

I also changed all RED action buttons to GREEN to signify a positive action as opposed to negative action.