Sorting posts & tags in ascending order, A-Z
Summary: If you sorted posts by "Title A-Z", it was sorting them "Z-A". Also if you sorted Tags by "Tag Name" it sorted them "Z-A" instead of "A-Z"

Test Plan: On the post list view, sort by "Title A-Z" and see the results. In workspace tab, clicl "Manage tags", sort by "Tag Name" and see results

Reviewers: shadowhand

Reviewed By: shadowhand

Subscribers: rjmackay

Differential Revision: https://phabricator.ushahidi.com/D79