Split JS router and controllers
Summary:
Trying out a cleaner structure for JS controllers and routers.
I've just split the settings route out for now as a simple test-case
but this makes even more sense for ie. messages or posts which have
multiple routes and controller functions.

- Move settings router and controller into 'SettingsApp'
- Move views and templates into settings subfolder
- Remove left over isMobile boilerplate

Test Plan: - Load UI and navigate to settings/site. See that this still works.

Reviewers: eyedol, shadowhand

Reviewed By: eyedol, shadowhand

Differential Revision: https://phabricator.ushahidi.com/D137