Ensure the post form is reloaded before display post create/edit
Summary:
Ensure the post form is reloaded before display post create/edit.
This is to make sure the form does not get out of sync after editing.

Test Plan:
Edit a form and its attributes.
Load post create and check the form is up to date

Edit a form and its attributes.
Load a post edit modal and check the form is up to date

Reviewers: aMoniker, shadowhand

Reviewed By: shadowhand

Maniphest Tasks: T681

Differential Revision: https://phabricator.ushahidi.com/D376