possibly fix premature oauth timeouts
Summary: Based on what Woody suggested, this could be a cookie issue. Last time it happened, I checked cookies and they were indeed missing, while the oauth_session_* tables in the database still contained my session keys. So, this sets the expire time of `authtoken` and `authrefresh` to the appropriate time value instead of to the default of 0, refs T903

Test Plan: log out and clear cookies, after logging in your cookies should have the proper expires values... it's difficult to catch this bug in action

Reviewers: rjmackay, shadowhand

Reviewed By: shadowhand

Subscribers: srutto

Maniphest Tasks: T903

Differential Revision: https://phabricator.ushahidi.com/D436