T1682 - API throws a 500 when adding a post to a set it is already part of
Summary:
After much jumping around between validators and factories, I decided I liked to check here best.

- if a post is already part of a collection, don't worry... be happy

Test Plan: using the list or detail views of post(s), add a post to a collection it already belongs to.  This should not cause any server errors to appear either in response, nor in the logs.  Likewise, adding a post to a collection it does not belong to should still work (i.e. the post and collection are associated in the table 'posts_sets'

Reviewers: rjmackay, lkamau, will, jasonmule

Subscribers: jasonmule, will, rjmackay, lkamau

Differential Revision: https://phabricator.ushahidi.com/D907