Fix API error adding a post to a set its already part of - T1682
Summary:
API throws a 500 when adding a post to a set it is already part of
If a post is already part of a collection, just pass through

Test Plan: using the list or detail views of post(s), add a post to a collection it already belongs to.  This should not cause any server errors to appear either in response, nor in the logs.  Likewise, adding a post to a collection it does not belong to should still work (i.e. the post and collection are associated in the table 'posts_sets'

Reviewers: lkamau, will, jasonmule, rjmackay

Reviewed By: rjmackay

Subscribers: jasonmule, will, rjmackay, lkamau

Differential Revision: https://phabricator.ushahidi.com/D907