refactor: Update CreateUsecase to allow default values on entities
Previously we created a blank entity and set state on it, this meant
that any values that had a default value *and* were marked as immutable
could not be set ie. `Post->type`. Instead CreateUsecase now creates
the entity with the full payload data, so the default are only applied
where the values doesn't appear in the payload, and the immutable
check works correctly because the value isn't yet set.